requirements.txt
# NeMo-Guardrails Test Suite Requirements

# Core dependencies
nemoguardrails>=0.15.0
python-dotenv>=1.0.0

# LLM Provider dependencies
google-generativeai>=0.8.0
langchain-google-genai>=2.1.0
langchain-openai>=0.1.0
langchain-core>=0.3.0

# System monitoring
psutil>=7.0.0

# Data handling
pydantic>=2.0.0
pandas>=1.4.0

# Async support
asyncio
nest-asyncio>=1.5.0

# HTTP client
httpx>=0.24.0
aiohttp>=3.10.0

# Utilities
rich>=13.5.0
typer>=0.8.0